fixup! move retrieval stats & tests to standalone files
